containers/netbox-mcp-server/requirements.txt
mcp[cli]
fastapi
uvicorn
httpx
sse-starlette
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
